refactor: Merge GroupsManager into GroupsSelect
